Safari-611.1.10/LayoutTests/http/tests/websocket/tests/hybi/interleaved-fragments.html

<!DOCTYPE HTML><!-- webkit-test-runner [ dumpJSConsoleLogInStdErr=true ] -->
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test whether WebSocket rejects interleaved fragmented frames.");

window.jsTestIsAsync = true;

var url = "ws://localhost:8880/websocket/tests/hybi/interleaved-fragments";
var ws = new WebSocket(url);
var closeEvent;

ws.onopen = function()
{
    debug("onopen() was called.");
};

ws.onmessage = function(event)
{
    var message = event.data;
    testFailed("onmessage() was called. (message = \"" + message + "\")");
};

ws.onclose = function(event)
{
    closeEvent = event;
    shouldBeFalse("closeEvent.wasClean");
    finishJSTest();
};

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
